Delete action
I have added a button which is linked to a delete action in the PostsController.

I used the setFlash method, like in the edit action, to show a pop up confirming the post was deleted. This time it is in red.

I've also changed the edit view so the Content input field is a text area.