posts filter action
Action responsible for filtering the posts by the author.

In the model for posts and authors, a relationship between the two ActiveRecords was formed with the code added here.

This allows me to access both database tables when displaying data to the user. They were linked via the ID and author_id columns.